Refactoring into classes
